make folding skipable

+++ benchmark.sh
@@ -29,31 +29,34 @@ fi
 
 
 #predict mfe structure and pairing probability matrices for all sequences using RNAfold and various SHAPE methods
-mkdir -p predictions
+if [[ "$*" != *--skipfolding* ]]
+then
+	mkdir -p predictions
 
-for rna in benchmarkdata/*.fa
-do
-	name=`echo $rna | sed 's/.[^.]*$//' | sed 's|benchmarkdata/||'`
+	for rna in benchmarkdata/*.fa
+	do
+		name=`echo $rna | sed 's/.[^.]*$//' | sed 's|benchmarkdata/||'`
 
-	echo "$name - RNAfold"
-	$rnafold -p --bppmThreshold=1e-15 < "benchmarkdata/$name.fa" > "predictions/$name.R.out"
-	mv *_dp.ps "predictions/$name.R.ps"
+		echo "$name - RNAfold"
+		$rnafold -p --bppmThreshold=1e-15 < "benchmarkdata/$name.fa" > "predictions/$name.R.out"
+		mv *_dp.ps "predictions/$name.R.ps"
 
-	echo "$name - Deigan"
-	$rnafold -p --bppmThreshold=1e-15 "--shape=benchmarkdata/$name.shape" --shapeMethod=D < "benchmarkdata/$name.fa" > "predictions/$name.D.out"
-	mv *_dp.ps "predictions/$name.D.ps"
+		echo "$name - Deigan"
+		$rnafold -p --bppmThreshold=1e-15 "--shape=benchmarkdata/$name.shape" --shapeMethod=D < "benchmarkdata/$name.fa" > "predictions/$name.D.out"
+		mv *_dp.ps "predictions/$name.D.ps"
 
-	echo "$name - Zarringhalam"
-	$rnafold -p --bppmThreshold=1e-15 "--shape=benchmarkdata/$name.shape" --shapeMethod=Z < "benchmarkdata/$name.fa" > "predictions/$name.Z.out"
-	mv *_dp.ps "predictions/$name.Z.ps"
+		echo "$name - Zarringhalam"
+		$rnafold -p --bppmThreshold=1e-15 "--shape=benchmarkdata/$name.shape" --shapeMethod=Z < "benchmarkdata/$name.fa" > "predictions/$name.Z.out"
+		mv *_dp.ps "predictions/$name.Z.ps"
 
-	echo "$name - Washietl"
-	$rnapvmin "benchmarkdata/$name.shape" < "benchmarkdata/$name.fa" > "predictions/$name.W.pv"
-	$rnafold -p --bppmThreshold=1e-15 "--shape=predictions/$name.W.pv" --shapeMethod=W < "benchmarkdata/$name.fa" > "predictions/$name.W.out"
-	mv *_dp.ps "predictions/$name.W.ps"
+		echo "$name - Washietl"
+		$rnapvmin "benchmarkdata/$name.shape" < "benchmarkdata/$name.fa" > "predictions/$name.W.pv"
+		$rnafold -p --bppmThreshold=1e-15 "--shape=predictions/$name.W.pv" --shapeMethod=W < "benchmarkdata/$name.fa" > "predictions/$name.W.out"
+		mv *_dp.ps "predictions/$name.W.ps"
 
-	rm *.ps
-done
+		rm *.ps
+	done
+fi
 
 
 #rate predictions and create plots
